Move window related built-in functions to evalwindow.c and buffer related built-in functions to evalbuffer.c
