+++ src/Make_morph.mak
@@ -41,8 +41,10 @@ SRC =	arabic.c						\
 	digraph.c						\
 	edit.c							\
 	eval.c							\
+	evalbuffer.c						\
 	evalfunc.c						\
 	evalvars.c						\
+	evalwindow.c						\
 	ex_cmds.c						\
 	ex_cmds2.c						\
 	ex_docmd.c						\
