+++ src/README.md
@@ -33,8 +33,10 @@ cmdhist.c	| command-line history
 debugger.c	| vim script debugger
 diff.c		| diff mode (vimdiff)
 eval.c		| expression evaluation
+evalbuffer.c	| buffer related built-in functions
 evalfunc.c	| built-in functions
 evalvars.c	| vim variables
+evalwindow.c	| window related built-in functions
 fileio.c	| reading and writing files
 filepath.c	| dealing with file names and paths
 findfile.c	| search for files in 'path'
