+++ src/proto/buffer.pro
@@ -29,6 +29,7 @@ char_u *buflist_nr2name(int n, int fullname, int helptail);
 void buflist_setfpos(buf_T *buf, win_T *win, linenr_T lnum, colnr_T col, int copy_options);
 void get_winopts(buf_T *buf);
 pos_T *buflist_findfpos(buf_T *buf);
+linenr_T buflist_findlnum(buf_T *buf);
 void buflist_list(exarg_T *eap);
 int buflist_name_nr(int fnum, char_u **fname, linenr_T *lnum);
 int setfname(buf_T *buf, char_u *ffname_arg, char_u *sfname_arg, int message);
@@ -69,24 +70,4 @@ void restore_win_for_buf(win_T *save_curwin, tabpage_T *save_curtab, bufref_T *s
 void set_buflisted(int on);
 int buf_contents_changed(buf_T *buf);
 void wipe_buffer(buf_T *buf, int aucmd);
-int set_ref_in_buffers(int copyID);
-buf_T * buflist_find_by_name(char_u *name, int curtab_only);
-buf_T * find_buffer(typval_T *avar);
-void f_append(typval_T *argvars, typval_T *rettv);
-void f_appendbufline(typval_T *argvars, typval_T *rettv);
-void f_bufadd(typval_T *argvars, typval_T *rettv);
-void f_bufexists(typval_T *argvars, typval_T *rettv);
-void f_buflisted(typval_T *argvars, typval_T *rettv);
-void f_bufload(typval_T *argvars, typval_T *rettv);
-void f_bufloaded(typval_T *argvars, typval_T *rettv);
-void f_bufname(typval_T *argvars, typval_T *rettv);
-void f_bufnr(typval_T *argvars, typval_T *rettv);
-void f_bufwinid(typval_T *argvars, typval_T *rettv);
-void f_bufwinnr(typval_T *argvars, typval_T *rettv);
-void f_deletebufline(typval_T *argvars, typval_T *rettv);
-void f_getbufinfo(typval_T *argvars, typval_T *rettv);
-void f_getbufline(typval_T *argvars, typval_T *rettv);
-void f_getline(typval_T *argvars, typval_T *rettv);
-void f_setbufline(typval_T *argvars, typval_T *rettv);
-void f_setline(typval_T *argvars, typval_T *rettv);
 /* vim: set ft=c : */
