+++ src/proto/window.pro
@@ -37,6 +37,8 @@ void tabpage_move(int nr);
 void win_goto(win_T *wp);
 win_T *win_find_nr(int winnr);
 tabpage_T *win_find_tabpage(win_T *win);
+win_T *win_vert_neighbor(tabpage_T *tp, win_T *wp, int up, long count);
+win_T *win_horz_neighbor(tabpage_T *tp, win_T *wp, int left, long count);
 void win_enter(win_T *wp, int undo_sync);
 win_T *buf_jump_open_win(buf_T *buf);
 win_T *buf_jump_open_tab(buf_T *buf);
@@ -58,7 +60,9 @@ void win_setminwidth(void);
 void win_drag_status_line(win_T *dragwin, int offset);
 void win_drag_vsep_line(win_T *dragwin, int offset);
 void set_fraction(win_T *wp);
+void win_new_height(win_T *wp, int height);
 void scroll_to_fraction(win_T *wp, int prev_height);
+void win_new_width(win_T *wp, int width);
 void win_comp_scroll(win_T *wp);
 void command_height(void);
 void last_status(int morewin);
@@ -79,38 +83,4 @@ int win_hasvertsplit(void);
 int get_win_number(win_T *wp, win_T *first_win);
 int get_tab_number(tabpage_T *tp);
 char *check_colorcolumn(win_T *wp);
-int win_getid(typval_T *argvars);
-int win_gotoid(typval_T *argvars);
-void win_id2tabwin(typval_T *argvars, list_T *list);
-win_T *win_id2wp(int id);
-win_T *win_id2wp_tp(int id, tabpage_T **tpp);
-int win_id2win(typval_T *argvars);
-void win_findbuf(typval_T *argvars, list_T *list);
-win_T *find_win_by_nr(typval_T *vp, tabpage_T *tp);
-win_T *find_win_by_nr_or_id(typval_T *vp);
-win_T *find_tabwin(typval_T *wvp, typval_T *tvp, tabpage_T **ptp);
-void f_gettabinfo(typval_T *argvars, typval_T *rettv);
-void f_getwininfo(typval_T *argvars, typval_T *rettv);
-void f_getwinpos(typval_T *argvars, typval_T *rettv);
-void f_getwinposx(typval_T *argvars, typval_T *rettv);
-void f_getwinposy(typval_T *argvars, typval_T *rettv);
-void f_tabpagenr(typval_T *argvars UNUSED, typval_T *rettv);
-void f_tabpagewinnr(typval_T *argvars, typval_T *rettv);
-void f_win_execute(typval_T *argvars, typval_T *rettv);
-void f_win_findbuf(typval_T *argvars, typval_T *rettv);
-void f_win_getid(typval_T *argvars, typval_T *rettv);
-void f_win_gotoid(typval_T *argvars, typval_T *rettv);
-void f_win_id2tabwin(typval_T *argvars, typval_T *rettv);
-void f_win_id2win(typval_T *argvars, typval_T *rettv);
-void f_win_screenpos(typval_T *argvars, typval_T *rettv);
-void f_winbufnr(typval_T *argvars, typval_T *rettv);
-void f_wincol(typval_T *argvars, typval_T *rettv);
-void f_winheight(typval_T *argvars, typval_T *rettv);
-void f_winlayout(typval_T *argvars, typval_T *rettv);
-void f_winline(typval_T *argvars, typval_T *rettv);
-void f_winnr(typval_T *argvars, typval_T *rettv);
-void f_winrestcmd(typval_T *argvars, typval_T *rettv);
-void f_winrestview(typval_T *argvars, typval_T *rettv);
-void f_winsaveview(typval_T *argvars, typval_T *rettv);
-void f_winwidth(typval_T *argvars, typval_T *rettv);
 /* vim: set ft=c : */
